Rename pip-requires to requirements.txt.
(cherry picked from commit ec9cf616904d92509b8b44662e7097352d894484)

+++ update.py
@@ -61,18 +61,31 @@ def _parse_reqs(filename):
     return reqs
 
 
-def _copy_requires(req, dest_dir):
+def _copy_requires(source_path, dest_dir):
     """Copy requirements files."""
 
-    dest_path = _dest_path(req, 'tools', dest_dir)
-    source_path = os.path.join('tools', req)
+    target_map = {
+        'requirements.txt': ('requirements.txt', 'tools/pip-requires'),
+        'test-requirements.txt': (
+            'test-requirements.txt', 'tools/test-requires'),
+    }
+    for dest in target_map[source_path]:
+        dest_path = os.path.join(dest_dir, dest)
+        if os.path.exists(dest_path):
+            break
+
+    # Catch the fall through
+    if not os.path.exists(dest_path):
+        # This can happen, we try all paths
+        return
 
     source_reqs = _parse_reqs(source_path)
     dest_reqs = _parse_reqs(dest_path)
     dest_keys = [key.lower() for key in dest_reqs.keys()]
     dest_keys.sort()
 
-    print "Syncing %s" % req
+    print "Syncing %s" % source_path
+
     with open(dest_path, 'w') as new_reqs:
         new_reqs.write("# This file is managed by openstack-depends\n")
         for old_require in dest_keys:
@@ -86,7 +99,7 @@ def _copy_requires(req, dest_dir):
                 new_reqs.write("%s\n" % dest_reqs[old_require])
 
 def main(argv):
-    for req in ('pip-requires', 'test-requires'):
+    for req in ('requirements.txt', 'test-requirements.txt'):
         _copy_requires(req, argv[0])
